requirements.txt
matplotlib==3.8.4
numpy==1.26.4
opencv_python_headless==4.9.0.80
Pillow==10.3.0
torch==2.2.2+cu118
torchvision==0.17.2+cu118
